@@ -1,9 +1,9 @@
 // https://stackoverflow.com/questions/42109813/node-js-environment-variables-and-heroku-deployment
 
-const dotenv = require('dotenv');
-const _ = require('lodash');
+import { config } from 'dotenv';
+import { each } from 'lodash';
 
-const result = dotenv.config();
+const result = config();
 
 let envs;
 
@@ -11,7 +11,7 @@ if (!('error' in result)) {
     envs = result.parsed;
 } else {
     envs = {};
-    _.each(process.env, (value, key) => envs[key] = value);
+    each(process.env, (value, key) => envs[key] = value);
 }
 
-module.exports = envs;
+export default envs;
