@@ -6,7 +6,7 @@
 // which each need 2 lines,
 // and Discord embeds are limited to 25 fields (which maths out to 24 fields here)
 
-const Discord = require('discord.js');
+import { MessageEmbed } from 'discord.js';
 
 const lister = {
     list_channels(message, dbchannels, title){
@@ -14,7 +14,7 @@ const lister = {
 
         let channels = message.guild.channels;
         //TODO: remove channel from database if it was deleted from the server
-        const embed = new Discord.MessageEmbed();
+        const embed = new MessageEmbed();
         embed.setTitle(title);
         const fields = [];
         for (var i = 0; i < dbchannels.length; i++){
@@ -31,4 +31,4 @@ const lister = {
     }
 };
 
-module.exports = lister;
+export default lister;
