@@ -1,10 +1,10 @@
-const Sequelize = require('sequelize');
+import Sequelize, { DataTypes } from 'sequelize';
 
-const envs = require('../config.js');
-const host = envs.DB_HOST;
-const database = envs.DB_NAME;
-const username = envs.DB_USER;
-const password = envs.DB_PASS;
+import { DB_HOST, DB_NAME, DB_USER, DB_PASS } from '../config.js';
+const host = DB_HOST;
+const database = DB_NAME;
+const username = DB_USER;
+const password = DB_PASS;
 
 const sequelize = new Sequelize(database, username, password, {
     host: host,
@@ -17,6 +17,6 @@ const sequelize = new Sequelize(database, username, password, {
     logging: false,
 });
 
-const Guilds = require('../models/guild_settings.js')(sequelize, Sequelize.DataTypes);
+const Guilds = require('../models/guild_settings.js').default(sequelize, DataTypes);
 
-module.exports = Guilds;
+export default Guilds;
