@@ -1,36 +1,38 @@
 "use strict";
 
 // TODO: consider moving anything that doesn't handle or rely on persistence
-// (i.e. sauce-related functions) to other files
+// (i.e. sauce-related functions) to other files and rename this file to datamanager.js or similar
 
 
 // TODO: Discord now allows multiple attachments per message sent by a non-bot user,
 // so now we have to figure out how to handle image dumps
+// we could queue x images from the dump and drip-feed them to SauceNao
+// keeps things from getting too spammy on Discord, SauceNao, etc.
 
-const Discord = require('discord.js');
-const https = require('https'); // TODO: either just use https, or just use got
-const got = require('got');
-const FileType = require('file-type');
-const parse = require('url-parse');
-const winston = require('winston');
-const { indexes, imageboards, videos, communities, communities2, domain_exclusions } = require('./saucenao_objects.js');
-const http_status_codes = require('../resources/http_status_codes.json');
-const envs = require('../config.js');
-const guildsDB = require('./db_objects.js');
+import { Collection, MessageEmbed } from 'discord.js';
+import { get } from 'https'; // TODO: either just use https, or just use got
+import { stream as _stream } from 'got';
+import { fromStream } from 'file-type';
+import parse from 'url-parse';
+import { createLogger, config, transports as _transports, format as _format } from 'winston';
+import { indexes, imageboards, videos, communities, communities2, domain_exclusions } from './saucenao_objects.js';
+import http_status_codes from '../resources/http_status_codes.json';
+import { BOT_COMMAND_PREFIX, BOT_TOKEN, SAUCENAO_API_KEY } from '../config.js';
+import guildsDB from './db_objects.js';
 
 class LoungeBot{
     constructor(){
-        this.logger = winston.createLogger({
-            levels: winston.config.syslog.levels,
+        this.logger = createLogger({
+            levels: config.syslog.levels,
             transports: [
-                new winston.transports.Console(),
+                new _transports.Console(),
                 //new winston.transports.File({ filename: 'log' }),
             ],
-            format: winston.format.combine(
-                winston.format.timestamp(),
-                winston.format.errors({ stack: true }),
+            format: _format.combine(
+                _format.timestamp(),
+                _format.errors({ stack: true }),
                 //winston.format.colorize(),
-                winston.format.printf(log => {
+                _format.printf(log => {
                     if (log.stack){
                         return `[${log.timestamp}] ${log.level.toUpperCase()}: ${log.message} - ${log.stack}`;
                     }
@@ -39,7 +41,7 @@ class LoungeBot{
             ),
         });
 
-        guildsDB.sync(); // make sure the database is initialized
+        guildsDB.sync(); // make sure the database is initialized (BROKEN?)
 
         // used for SauceNAO backoff (measured in ms)
         this._saucenaoTimeout = {
@@ -47,10 +49,10 @@ class LoungeBot{
             short: true
         };
 
-        this.command_prefix = envs.BOT_COMMAND_PREFIX;
-        this.token = envs.BOT_TOKEN;
-        this.api_key = envs.SAUCENAO_API_KEY;
-        this.guilds = new Discord.Collection();
+        this.command_prefix = BOT_COMMAND_PREFIX;
+        this.token = BOT_TOKEN;
+        this.api_key = SAUCENAO_API_KEY;
+        this.guilds = new Collection();
     }
 
     //------ DATABASE METHODS ------
@@ -264,7 +266,7 @@ class LoungeBot{
                 break; // bail out
             }
 
-            const embed = new Discord.MessageEmbed();
+            const embed = new MessageEmbed();
             // add result header data
             if (manually_invoked){ embed.setTitle("Result"); }
             else { embed.setTitle("Best Guess for This Image"); }
@@ -510,7 +512,7 @@ class LoungeBot{
         }*/
 
         if (this._saucenaoTimeout.duration - Date.now() > 0) {
-            const embed = new Discord.MessageEmbed();
+            const embed = new MessageEmbed();
             if (this._saucenaoTimeout.short) {
                 embed.setDescription(`Too many searches within 30 seconds. Try again in ${humanReadableDuration(this._saucenaoTimeout.duration)}.`);
                 //return message.reply({ allowedMentions: { repliedUser: false }, embeds: [embed] });
@@ -592,8 +594,8 @@ class LoungeBot{
 
         // save SauceNAO some trouble and check if image URL actually leads to an image
         try {
-            const stream = got.stream(imgurl);
-            const mimetype = await FileType.fromStream(stream);
+            const stream = _stream(imgurl);
+            const mimetype = await fromStream(stream);
 
             const re = /^image\/.+$/;
             if (!mimetype || !re.test(mimetype.mime)) {
@@ -643,7 +645,7 @@ class LoungeBot{
         + `&url=${imgurl}`;
 
         // https://www.twilio.com/blog/2017/08/http-requests-in-node-js.html
-        https.get(url, (resp) => {
+        get(url, (resp) => {
             const status_msg = http_status_codes;
             if (resp.statusCode !== 200){
                 if (resp.statusCode === 403){
@@ -654,7 +656,7 @@ class LoungeBot{
                     const error_msg = `SauceNAO returned **${resp.statusCode} ${status_msg[String(resp.statusCode)]}**`;
                     let nsfw_warning = "";
                     if (!message.channel.nsfw){ nsfw_warning = "(may or may not be NSFW)"; }
-                    const embed = new Discord.MessageEmbed();
+                    const embed = new MessageEmbed();
                     embed.setTitle("Search Failed");
                     embed.setDescription(error_msg);
                     embed.addField("Fallback URL", `[SauceNAO results page](${fallback_url}) ${nsfw_warning}`);
@@ -706,7 +708,7 @@ class LoungeBot{
                     if (!results){
                         const user_error_msg = "SauceNAO API Error. Try again.";
                         if (options.manually_invoked){
-                            return message.reply({ embeds: [new Discord.MessageEmbed().setDescription(user_error_msg)] });
+                            return message.reply({ embeds: [new MessageEmbed().setDescription(user_error_msg)] });
                         }
                         return;
                     }
@@ -716,7 +718,7 @@ class LoungeBot{
                         const error_msg = `Search failed with message: \`\`${header.message}\`\``;
                         this.logger.warning(error_msg);
                         if (options.manually_invoked){
-                            return message.reply({ embeds: [new Discord.MessageEmbed().setDescription(error_msg)] });
+                            return message.reply({ embeds: [new MessageEmbed().setDescription(error_msg)] });
                         }
                         return;
                     }
@@ -729,7 +731,7 @@ class LoungeBot{
                     const error_msg = "SauceNAO reported successful search but returned no results.";
                     this.logger.error(error_msg);
                     if (options.manually_invoked){
-                        return message.reply({ embeds: [new Discord.MessageEmbed().setDescription(error_msg)] });
+                        return message.reply({ embeds: [new MessageEmbed().setDescription(error_msg)] });
                     }
                     return;
                 }
@@ -768,4 +770,4 @@ class LoungeBot{
     // --------------------------
 }
 
-module.exports = LoungeBot;
+export default LoungeBot;
