@@ -80,4 +80,4 @@ const domain_exclusions = [
 ];
 
 
-module.exports = { indexes, imageboards, videos, communities, communities2, domain_exclusions }
+export default { indexes, imageboards, videos, communities, communities2, domain_exclusions }
